+++ src/model/model.ts
@@ -1,11 +1,16 @@
+import { RepositoryUtils } from "../utils/repositoryUtils";
+
 class MongoModel {
 
-    private static schema : any = {};
+    private static schema: any = {};
 
-    setModel(inputSchema : any) {
-        MongoModel.schema = Object.assign(MongoModel.schema, inputSchema);
+    setModel(inputSchema: any) {
+        if (RepositoryUtils.isValidQueryUpdateObject(inputSchema))
+            MongoModel.schema = Object.assign(MongoModel.schema, inputSchema);
     }
-    getSchemaJson(schemaName : string) {
+    getSchemaJson(schemaName: string) {
+        if(!schemaName || typeof schemaName != "string")
+            return undefined;
         return MongoModel.schema[schemaName];
     }
 }