+++ test/crudOperations.test.js
@@ -11,14 +11,14 @@ const config = require('./config');
 const OnTheFlyRepositoryFactory = require('../lib/index').OnTheFlyRepositoryFactory;
 
 // test cases that must be passed before
-require('./connection.test');
+require('./data.test');
 
 OnTheFlyRepositoryFactory.loadModel(config.TestSchema1ModelLoading)
 
 describe("CRUD operations tests", function () {
     this.timeout(20000);
 
-    it("CRUD operations for a singleton repository", async function () {
+    it("CRUD operations for a prototype repository", async function () {
         let ret;
         let err;
         try {
@@ -68,5 +68,48 @@ describe("CRUD operations tests", function () {
              assert.equal(err, null || undefined)
          }
      })
+
+     it("CRUD operations Insert Many and Find All for a prototype repository", async function () {
+        let ret;
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigPrototypeTestSchema1).insert([config.InsertQueryTestSchema1, config.InsertQueryTestSchema1Many1]);
+            assert.notEqual(ret,  null || undefined)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigPrototypeTestSchema1).find({query : {}});
+            assert.notEqual(ret,  null || undefined)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigPrototypeTestSchema1).remove(config.InsertQueryTestSchema1);
+            assert.notEqual(ret,  null || undefined)
+            assert.equal(ret.ok, 1)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigPrototypeTestSchema1).remove(config.InsertQueryTestSchema1Many1);
+            assert.notEqual(ret,  null || undefined)
+            assert.equal(ret.ok, 1)
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.equal(err, null || undefined)
+        }
+    })
+
+    it("CRUD operations Insert Many and Find All for a singleton repository", async function () {
+        let ret;
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigSingletonTestSchema1).insert([config.InsertQueryTestSchema1Many2, config.InsertQueryTestSchema1Many3]);
+            assert.notEqual(ret,  null || undefined)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigSingletonTestSchema1).find({query : {}});
+            assert.notEqual(ret,  null || undefined)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigSingletonTestSchema1).remove(config.InsertQueryTestSchema1Many2);
+            assert.notEqual(ret,  null || undefined)
+            assert.equal(ret.ok, 1)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigSingletonTestSchema1).remove(config.InsertQueryTestSchema1Many3);
+            assert.notEqual(ret,  null || undefined)
+            assert.equal(ret.ok, 1)
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.OkConfigSingletonTestSchema1).closeSingletonConnection();
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.equal(err, null || undefined)
+        }
+    })
 
 })