+++ test/secret.test.js
@@ -14,8 +14,179 @@ require('./advancedConnectionTest.test');
 
 OnTheFlyRepositoryFactory.loadModel(config.SecretSchemaModelLoading)
 
-describe("Secret Repository tests", function () {
+describe("Secret Repository field tests", function () {
     this.timeout(20000);
+
+    it("The Secret Repository must return an error for wrong type paraamters field (string)", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "secretSchema",
+                singleton: false,
+                secret: true,
+                parameters: "secret"
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for wrong type paraamters field (json) ", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "secretSchema",
+                singleton: false,
+                secret: true,
+                parameters: {}
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for wrong type paraamters field (type) ", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "secretSchema",
+                singleton: false,
+                secret: true,
+                parameters: String
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+
+    it("The Secret Repository must return an error for wrong type paraamters field (array of int) ", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "secretSchema",
+                singleton: false,
+                secret: true,
+                parameters: [1,2,3]
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for spaces url", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "  ",
+                schemaName: "secretSchema",
+                singleton: false,
+                secret: true,
+                parameters: ["secret"]
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for spaces schemaName", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "  ",
+                singleton: false,
+                secret: true,
+                parameters: ["secret"]
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for not loaded schema", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository({
+                url: "url",
+                schemaName: "XXXX",
+                singleton: false,
+                secret: true,
+                parameters: ["secret"]
+            }).find(config.FindQuerySecretSchema);
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for empty query parameter", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.ConfigSecretSchema).find({});
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for wrong query parameter", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.ConfigSecretSchema).find({ query : String});
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for empty update parameter in update operation", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.ConfigSecretSchema).update({ query : {}});
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+    it("The Secret Repository must return an error for wrong update parameter in update operation", async function () {
+        let err;
+        try {
+            ret = await OnTheFlyRepositoryFactory.generateOnTheFlyRepository(config.ConfigSecretSchema).update({ query : {}, update : String});
+        } catch (e) {
+            err = e;
+        } finally {
+            assert.notEqual(err, null || undefined)
+        }
+    })
+
+})
+
+describe("Secret Repository operation tests", function () {
+    this.timeout(20000);
+
+
 
     it("The crypted information must be encrypted in the reading part", async function () {
         let ret;